pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dot-PE"
version = "0.1.0"
authors = [
    { name = "Jonathan Mushkin", email = "yonatanmushkin@gmail.com" },
]
description = "A short description of your project"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "GPL-3.0"}
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "cogwheel-pe>=1.2.0",
    "pandas",
    "scipy",
    "numpy",
    "torch",
    "matplotlib",
    "tqdm",
    "lal",
    "lalsimulation"
]

[tool.setuptools.packages.find]
include = ["dot_pe*"]
exclude = ["notebooks*", "tests*", "scripts*"] 
